+++ src/main/java/chess/model/dao/TurnDao.java
@@ -1,27 +1,15 @@
 package chess.model.dao;
 
+import chess.utils.DBConnector;
+
 import java.sql.*;
 
 public class TurnDao {
-    private static final String URL = "jdbc:mysql://localhost:3306/chess";
-    private static final String USER = "user";
-    private static final String PASSWORD = "password";
-
-    public static Connection getConnection() {
-        Connection conn = null;
-        try {
-            conn = DriverManager.getConnection(URL, USER, PASSWORD);
-        } catch (final Exception e) {
-            System.out.println("커넥션 연결에 실패하였습니다.");
-            e.printStackTrace();
-        }
-        return conn;
-    }
+    private static final Connection connection = DBConnector.getConnection();
 
     public void init() {
         String query = "insert into turns (turn) values (?)";
-
-        try (PreparedStatement preparedStatement = getConnection().prepareStatement(query)) {
+        try (PreparedStatement preparedStatement = connection.prepareStatement(query)) {
             preparedStatement.setString(1, "WHITE");
             preparedStatement.executeUpdate();
         } catch (SQLException throwables) {
@@ -32,11 +20,10 @@ public class TurnDao {
     public String findOne() {
         String query = "select turn from turns limit 1";
         String turn = "";
-        try {
-            PreparedStatement preparedStatement = getConnection().prepareStatement(query);
+        try (PreparedStatement preparedStatement = connection.prepareStatement(query)) {
             ResultSet resultSet = preparedStatement.executeQuery();
             resultSet.next();
-             turn = resultSet.getString("turn");
+            turn = resultSet.getString("turn");
         } catch (SQLException throwables) {
             throwables.printStackTrace();
         }
@@ -45,8 +32,7 @@ public class TurnDao {
 
     public void update(String nextTurn) {
         String query = "UPDATE turns SET turn = (?)";
-        try {
-            PreparedStatement preparedStatement = getConnection().prepareStatement(query);
+        try (PreparedStatement preparedStatement = connection.prepareStatement(query)) {
             preparedStatement.setString(1, nextTurn);
             preparedStatement.executeUpdate();
         } catch (SQLException throwables) {
@@ -56,8 +42,7 @@ public class TurnDao {
 
     public void deleteAll() {
         String query = "DELETE FROM turns";
-        try {
-            PreparedStatement preparedStatement = getConnection().prepareStatement(query);
+        try (PreparedStatement preparedStatement = connection.prepareStatement(query)) {
             preparedStatement.executeUpdate();
         } catch (SQLException throwables) {
             throwables.printStackTrace();
